Improved DB Schema + Comments (#6)
* fixed consistent comment indices

* actually fixed comment indices

* database schema revamp with improved constraints

* fix unnecessary negative comment and link index

* format and remove comment

* bug fix in comment deletion

+++ src/main/scala/muse/domain/mutate/CreateComment.scala
@@ -9,7 +9,7 @@ import java.util.UUID
 final case class CreateComment(
     reviewId: UUID,
     // If none, then it is root comment.
-    parentCommentId: Option[Int],
+    parentCommentId: Option[Long],
     comment: String,
     entities: List[ReviewEntityInput],
     // If none, will be added to the end of the list.
